+++ eyes_common/applitools/common/app_output.py
@@ -7,15 +7,15 @@ from .utils.json_utils import JsonInclude
 if TYPE_CHECKING:
     from typing import Optional, Text
 
-    from .geometry import RectangleSize
+    from .geometry import Point, RectangleSize
 
 __all__ = ("AppOutput",)
 
 
 @attr.s
 class AppOutput(object):
-    title = attr.ib(metadata={JsonInclude.THIS: True})  # type: Text
-    screenshot_bytes = attr.ib(repr=False)  # type: Optional[bytes]
+    title = attr.ib(default=None, metadata={JsonInclude.NON_NONE: True})  # type: Text
+    screenshot_bytes = attr.ib(default=None, repr=False)  # type: Optional[bytes]
     screenshot_url = attr.ib(
         default=None, metadata={JsonInclude.NON_NONE: True}
     )  # type: Optional[Text]
@@ -25,3 +25,6 @@ class AppOutput(object):
     viewport = attr.ib(
         default=None, metadata={JsonInclude.NON_NONE: True}
     )  # type: Optional[RectangleSize]
+    location = attr.ib(
+        default=None, metadata={JsonInclude.NON_NONE: True}
+    )  # type: Optional[Point]