+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [vNext]
+## [4.16.0] - 2021-02-05
 ### Added
 - Add extract_text_region for OCR support [Trello 2440](https://trello.com/c/FLGKnqIS)
 ### Fixed